feat: send turn complete event to players when game abort
